Add tests in CollectionsTest and StringsTest. Remove string tests from UtilsTest
